strategy refactoring
